requirements.txt
# Core dependencies (Python 3.11 compatible)
typing-extensions>=4.8.0
numpy>=1.21.0,<1.27.0
pandas>=1.5.0,<2.3.0
scipy>=1.9.0
scikit-learn>=1.1.0,<1.4.0
joblib>=1.3.0

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0,<2.1.0

# Async processing
aiohttp>=3.8.5
asyncpg>=0.28.0
anyio>=4.0.0

# Machine Learning 
xgboost>=1.6.0
lightgbm>=3.3.0
torch>=2.0.0
#tensorflow>=2.12.0,<2.20.0
tensorflow[and-cuda]

# Data processing
pandas-ta==0.3.14b0
yfinance>=0.2.18
beautifulsoup4>=4.12.0

# Visualization
plotly>=5.15.0
dash>=3.0.0
dash-bootstrap-components>=1.4.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-core>=2.0.0

# Monitoring
prometheus-client>=0.17.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Additional utilities
requests>=2.31.0
